Rename the HomePageHandler
This change isn't necessary, but helpful so that the handler and its
factory are named after the handler's new intent. The change in
functioning of the handler will come later.
