+++ config/routes.php
@@ -13,7 +13,7 @@ use Psr\Container\ContainerInterface;
  *
  * Setup routes with a single request method:
  *
- * $app->get('/', App\Handler\HomePageHandler::class, 'home');
+ * $app->get('/', App\Handler\CodeRequestFormPageHandler::class, 'home');
  * $app->post('/album', App\Handler\AlbumCreateHandler::class, 'album.create');
  * $app->put('/album/{id:\d+}', App\Handler\AlbumUpdateHandler::class, 'album.put');
  * $app->patch('/album/{id:\d+}', App\Handler\AlbumUpdateHandler::class, 'album.patch');
@@ -38,6 +38,6 @@ use Psr\Container\ContainerInterface;
  */
 
 return static function (Application $app, MiddlewareFactory $factory, ContainerInterface $container): void {
-    $app->get('/', App\Handler\HomePageHandler::class, 'home');
+    $app->get('/', App\Handler\CodeRequestFormPageHandler::class, 'code.request.form');
     $app->get('/api/ping', App\Handler\PingHandler::class, 'api.ping');
 };
