+++ src/App/src/ConfigProvider.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace App;
 
+use Laminas\ServiceManager\AbstractFactory\ReflectionBasedAbstractFactory;
+
 /**
  * The configuration provider for the App module
  *
@@ -35,7 +37,7 @@ class ConfigProvider
                 Handler\PingHandler::class => Handler\PingHandler::class,
             ],
             'factories'  => [
-                Handler\HomePageHandler::class => Handler\HomePageHandlerFactory::class,
+                Handler\CodeRequestFormPageHandler::class        => ReflectionBasedAbstractFactory::class,
             ],
         ];
     }
